product update and delete routes

+++ app/forms/product_create.py
@@ -1,11 +1,11 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField,SelectField,BooleanField, DecimalField
+from wtforms import StringField, SubmitField,SelectField,BooleanField, FloatField
 from wtforms.validators import DataRequired
 
 
 class CreateProductForm(FlaskForm):
   name=StringField('Name', validators=[DataRequired()])
-  price=DecimalField("Price", validators=[DataRequired()])
+  price=FloatField("Price", validators=[DataRequired()])
   description = StringField('Description', validators=[DataRequired()])
   category = SelectField("Category", choices=[("Clothing", "Clothing"),("Jewelry", "Jewelry"),("Home & Living", "Home & Living"),("Craft Supplies", "Craft Supplies")])
   return_accepted = BooleanField("Is refundable")
